+++ pybamm/models/submodels/interface/sei/ec_reaction_limited.py
@@ -119,13 +119,20 @@ class EcReactionLimited(BaseModel):
         if self.reaction_loc == "x-average":
             L_sei = variables["X-averaged outer SEI thickness"]
             j_sei = variables["X-averaged outer SEI interfacial current density"]
+            # Note a is dimensionless (has a constant value of 1 if the surface
+            # area does not change)
+            a = variables["X-averaged negative electrode surface area to volume ratio"]
         else:
             L_sei = variables["Outer SEI thickness"]
             j_sei = variables["Outer SEI interfacial current density"]
+            if self.reaction_loc == "interface":
+                a = 1
+            else:
+                a = variables["Negative electrode surface area to volume ratio"]
 
         Gamma_SEI = self.param.Gamma_SEI
 
-        self.rhs = {L_sei: -Gamma_SEI * j_sei / 2}
+        self.rhs = {L_sei: -Gamma_SEI * a * j_sei / 2}
 
     def set_initial_conditions(self, variables):
         if self.reaction_loc == "x-average":